+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rasky/utilities/wrappers/WrappedServo.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.rasky.utilities;
+package org.firstinspires.ftc.teamcode.rasky.utilities.wrappers;
 
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.HardwareMap;
